Better control of which site test runs again. Appserver will not be started if URL is specified.

+++ test/test_all.sh
@@ -1,12 +1,18 @@
 #!/bin/bash
 cd $(dirname "$0")
 
-echo "Starting the dev appserver..."
-dev_appserver.py .. &> appserver.log &
-sleep 3
+if [ -z $REFTAG_SITE ]
+then
+    echo "Starting the dev appserver..."
+    dev_appserver.py .. &> appserver.log &
+    sleep 3
+fi
 
 echo "Running the test..."
 python -m unittest test_with_browser test_jsonp_api
 
-echo "Stopping the dev appserver..."
-kill $!
+if [ -z $REFTAG_SITE ]
+then
+    echo "Stopping the dev appserver..."
+    kill $!
+fi
